requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
nltk==3.8.1
numpy==1.24.1
oauthlib==3.2.2
packaging==22.0
pandas==1.5.2
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
textblob==0.17.1
tqdm==4.64.1
tweepy==4.12.1
urllib3==1.26.13
wordcloud==1.8.2.2
